+++ src/modules/players/domain/index.ts
@@ -1,9 +1,0 @@
-export { Player, type PlayerProps, type PlayerIdProps } from './Player.entity';
-export type { IPlayerRepository } from './Player.repository';
-export { EmailAlreadyInUseError } from './errors';
-export { Email, PhoneNumber, Birthdate, PlayerId } from './value-objects';
-export {
-  PlayerCategory,
-  isPlayerCategory,
-  parsePlayerCategory,
-} from './PlayerCategory';